requirements.txt
pytest
pylint
ipdb
sentencepiece
transformers
easyocr
sacremoses
pillow
<<<<<<< Updated upstream
=======
pytesseract
>>>>>>> Stashed changes

# API
fastapi
pytz
uvicorn
